Get rid of unused assets and get rid of unused ‘dumb_components’ folder in shared.
