+++ src/admin/containers/add_member.js
@@ -4,7 +4,7 @@ const { connect } = require('react-redux')
 const { reduxForm } = require('redux-form')
 const { pick } = require('ramda')
 
-const MemberFields = require('../../shared/dumb_components/member_fields.js')
+const MemberFields = require('../../shared/components/member_fields.js')
 const { fields, validate, new_required, read_only } = require('../../shared/form_fields/member.js')
 const { create_member } = require('../../shared/redux/modules/member.js')
 const { send_welcome } = require('../redux/modules/email/reducer.js')
