+++ src/open/app.js
@@ -6,8 +6,8 @@ var App = React.createClass({
 
   render: function () {
     return (
-	<div>
-	{this.props.children}
+      <div>
+        {this.props.children}
       </div>
     )
   }
