+++ src/shared/components/change_password_form.js
