+++ src/user/components/change_password.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import ChangePassword from '../../shared/components/change_password.js'
+import ChangePassword from '../../shared/containers/change_password.js'
 
 export default () =>
   <div className='change-password-section'>
